backend/pyproject.toml
[project]
name = "ember-api"
version = "0.1.0"
description = "Ember AI-Powered Romantasy Ghostwriting API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "anthropic>=0.18.0",
    "pgvector>=0.2.4",
    "python-docx>=1.1.0",
    "ebooklib>=0.18",
    "stripe>=8.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "black", "ruff"]
